requirements.txt
flask>=3.0
authlib>=1.3
anthropic>=0.40
pyjwt>=2.8
python-dotenv>=1.0
gunicorn>=22.0
